+++ tests/unit/base/test_pipelines_builders_find_device_max_batch_size.py
@@ -19,7 +19,7 @@ from model_navigator.configuration import (
     TensorRTPrecisionMode,
 )
 from model_navigator.configuration.common_config import CommonConfig
-from model_navigator.configuration.constants import DEFAULT_MAX_WORKSPACE_SIZE
+from model_navigator.configuration.constants import DEFAULT_MAX_WORKSPACE_SIZE, DEFAULT_ONNX_OPSET
 from model_navigator.configuration.model.model_config import (
     ONNXModelConfig,
     TensorFlowSavedModelConfig,
@@ -65,7 +65,11 @@ def test_find_device_max_batch_size_builder_return_execution_unit_when_torch_fra
                 inference_mode=True,
             )
         ],
-        Format.ONNX: [ONNXModelConfig(opset=17, dynamic_axes={}, dynamo_export=False, graph_surgeon_optimization=True)],
+        Format.ONNX: [
+            ONNXModelConfig(
+                opset=DEFAULT_ONNX_OPSET, dynamic_axes={}, dynamo_export=False, graph_surgeon_optimization=True
+            )
+        ],
         Format.TENSORRT: [
             TensorRTModelConfig(
                 precision=TensorRTPrecision.FP16,
@@ -103,7 +107,11 @@ def test_find_device_max_batch_size_builder_return_execution_unit_when_tensorflo
 
     models_config = {
         Format.TF_SAVEDMODEL: [TensorFlowSavedModelConfig(enable_xla=False, jit_compile=False)],
-        Format.ONNX: [ONNXModelConfig(opset=17, dynamic_axes={}, dynamo_export=False, graph_surgeon_optimization=True)],
+        Format.ONNX: [
+            ONNXModelConfig(
+                opset=DEFAULT_ONNX_OPSET, dynamic_axes={}, dynamo_export=False, graph_surgeon_optimization=True
+            )
+        ],
         Format.TENSORRT: [
             TensorRTModelConfig(
                 precision=TensorRTPrecision.FP16,
@@ -142,7 +150,11 @@ def test_find_device_max_batch_size_builder_return_execution_unit_when_jax_frame
 
     models_config = {
         Format.TF_SAVEDMODEL: [TensorFlowSavedModelConfig(enable_xla=True, jit_compile=True)],
-        Format.ONNX: [ONNXModelConfig(opset=17, dynamic_axes={}, dynamo_export=False, graph_surgeon_optimization=True)],
+        Format.ONNX: [
+            ONNXModelConfig(
+                opset=DEFAULT_ONNX_OPSET, dynamic_axes={}, dynamo_export=False, graph_surgeon_optimization=True
+            )
+        ],
         Format.TENSORRT: [
             TensorRTModelConfig(
                 precision=TensorRTPrecision.FP16,
@@ -180,7 +192,11 @@ def test_find_device_max_batch_size_builder_return_execution_unit_when_onnx_fram
     )
 
     models_config = {
-        Format.ONNX: [ONNXModelConfig(opset=17, dynamic_axes={}, dynamo_export=False, graph_surgeon_optimization=True)],
+        Format.ONNX: [
+            ONNXModelConfig(
+                opset=DEFAULT_ONNX_OPSET, dynamic_axes={}, dynamo_export=False, graph_surgeon_optimization=True
+            )
+        ],
         Format.TENSORRT: [
             TensorRTModelConfig(
                 precision=TensorRTPrecision.FP16,