+++ tests/unit/torch/test_pipelines_builders_graph_surgeon.py
@@ -12,7 +12,7 @@
 # See the License for the specific language governing permissions and
 
 from model_navigator.commands.optimize.graph_surgeon import GraphSurgeonOptimize
-from model_navigator.configuration import DeviceKind, Format, JitType, OptimizationProfile
+from model_navigator.configuration import DEFAULT_ONNX_OPSET, DeviceKind, Format, JitType, OptimizationProfile
 from model_navigator.configuration.common_config import CommonConfig
 from model_navigator.configuration.model.model_config import ONNXModelConfig, TorchScriptModelConfig
 from model_navigator.frameworks import Framework
@@ -32,7 +32,11 @@ def test_torch_export_builder_return_graph_surgeon_optimization_when_enabled():
     )
 
     models_config = {
-        Format.ONNX: [ONNXModelConfig(opset=17, dynamic_axes={}, dynamo_export=False, graph_surgeon_optimization=True)],
+        Format.ONNX: [
+            ONNXModelConfig(
+                opset=DEFAULT_ONNX_OPSET, dynamic_axes={}, dynamo_export=False, graph_surgeon_optimization=True
+            )
+        ],
     }
     pipeline = torch_export_builder(config=config, models_config=models_config)
     assert len(pipeline.execution_units) == 2
@@ -53,7 +57,9 @@ def test_torch_export_builder_does_not_return_graph_surgeon_optimization_when_di
 
     models_config = {
         Format.ONNX: [
-            ONNXModelConfig(opset=17, dynamic_axes={}, dynamo_export=False, graph_surgeon_optimization=False)
+            ONNXModelConfig(
+                opset=DEFAULT_ONNX_OPSET, dynamic_axes={}, dynamo_export=False, graph_surgeon_optimization=False
+            )
         ],
     }
     pipeline = torch_export_builder(config=config, models_config=models_config)
@@ -75,7 +81,7 @@ def test_torch_conversion_builder_return_graph_surgeon_optimization_when_enabled
     models_config = {
         Format.ONNX: [
             ONNXModelConfig(
-                opset=17,
+                opset=DEFAULT_ONNX_OPSET,
                 dynamic_axes={},
                 dynamo_export=False,
                 graph_surgeon_optimization=True,
@@ -105,7 +111,7 @@ def test_torch_conversion_builder_does_not_return_graph_surgeon_optimization_whe
     models_config = {
         Format.ONNX: [
             ONNXModelConfig(
-                opset=17,
+                opset=DEFAULT_ONNX_OPSET,
                 dynamic_axes={},
                 dynamo_export=False,
                 graph_surgeon_optimization=False,